requirements.txt
dspy-ai
exa_py==1.1.7
google_api_python_client==2.108.0
google_auth_oauthlib==1.1.0
protobuf==5.28.2
python-dotenv==1.0.1
youtube_transcript_api==0.5.0
google-generativeai==0.8.1
groq==0.11.0